Break up branch_length_estimator.py into smaller modules
